Add instructions issuance logic

+++ app/src/main/java/com/husseinelfeky/tomasulosimulator/model/simulation/base/Buffer.kt
@@ -4,7 +4,7 @@ import com.husseinelfeky.tomasulosimulator.model.simulation.general.Tag
 
 abstract class Buffer(
     override val tag: Tag,
-    open val address: Double?,
+    open val address: Int?,
     override val isBusy: Boolean,
     override val remainingCycles: Int?
 ) : RowItem(tag, isBusy, remainingCycles)
